+++ libs/cmusy-ui/src/index.ts
@@ -10,7 +10,7 @@ export { default as Typography } from './lib/typography';
 export { default as Container } from './lib/container';
 export { default as Pill } from './lib/pill';
 export { default as Rating } from './lib/rating';
-
+export * from './lib/datepicker';
 export * from './lib/theme';
 export * from './lib/tabs';
 export * from './lib/card';